Improvements to the way modes are created
Managing modes is a big simpler now; instead of a secondary mode
deviation, the mode is set on the buffer, so it is clear which buffer uses
which mode, and ownership of one-time modes is clearer.

Some more work on the repl, and an example lisp interpreter for the
demo app to use; shows how you hook into the generic repl mode.  This
is experimental for now, and supplying a repl will be optional.  The end
user just supplies a <string(string)> std::function which takes
something to evaluate and returns the result.  The editor does the job
of being a 'readline' clone.