Cursor Line and Window Drag
Fix dragging the window by clicking on it; remove the Invisible Button
and detect window clicks that are not handled.  Still need to add mouse
selection of lines for non-keyboard folks.

Add cursor_line_solid variable to editor; for either a background
lighter fill or a background line above/below selection.