Keymapper
In order for different modes to handle keys differently, and eventually
to allow the user to tweak key shortcuts, I need a keyboard mapping.
This starts that process.  A keystroke combination is stored in a tree
which is analyzed as the user types commands (in this way it can detect
'unfinished' commands, and handle combinations that should happen after
a given time if no more keys are pressed, etc.)
This CL has a couple of keystrokes added to the map on the vim side; but
I will add more as I go.