Mode fixing
This branch definately shouldn't be used right now, as I'm reworking
things and breaking others.

This work in progress is slowly generalizing the keyboard input commands
so that they can go through a mapping phase depending on modes.  It is
also simplifying and fixing the undo/redo and dot operations to always
behave as expected.  This cleanup feels good, but needs a few more CLs
to get it all done.  Standard mode is broken in the mean time, and dot
mode in vim is not working til my next update.